basebackup: support encrypting and compressing basebackups on the fly
Previously we first stored the basebackup onto disk after which we
compressend and encrypted it. This meant we'd need roughly 2.5x
diskspace (1 copy in the running db, 1 as a cleartext tempfile and
~0.5x for the compressed encrypted copy).

This drops the requirement to roughly ~1.5x (1 running copy and a
second temporary compressed and encrypted one. Eventually we should
just create a stacked encryptor/compressor that would take a fp as
an input and give out a stacked file handle that you could read from
to write into an object storage.