storage config: simplify get_object_storage_config and get_transfer apis
get_object_storage_config now always returns just a dict with storage_type
in it which can be passed to get_transfer.  get_transfer allows storage_type
to be passed as a keyword argument or inside storage_config, but doesn't
pass it on to the drivers which don't expect to see it.