swift: cut down logging for gets failing on 404
Previously the swift client created a couple of pages of logs (at INFO
level) for every GET to a nonexistent file as it logged a full traceback as
well as some made up curl commands to try to perform the same request that
failed.

Set Swift's log level to WARNING and filter its exception logging a bit.  We
still log a lot, but this makes it somewhat manageable.

We'll probably need to do something to cut down the amount of logs we
generate in cases where a PG cluster is set up as a standby only connected
to the WAL archive.  PG keeps asking us for files that don't exist and we
keep logging 4-5 lines for every such request.