azure: Make Azure support work with latest azure-storage-python
Azure-storage-python had in the meanwhile changed the names of
multiple methods and changed the behavior around listing of files.

Azure no longer accepts hyphen's '-', we replace them in metadata
now with an underscore '_' and then replace them back when doing
blob listings.

Also improve error handling so missing files are handled gracefully,
and add Azure to the list of tested storage backends. The storage
completes the tests successfully after this commit.