google: compatibility with recent oauth2client versions
The oauth2client made the previously _ServiceAccountCredentials private
class public some time ago, but now the new class has changed its __init__
method's required arguments.  Instead of adapting for it and trying to
support two variants of it, we can use the from_json_keyfile_dict
classmethod that was added when it was made public.

We'll still support the old _ServiceAccountCredentials way for a while.