dates: start using a tz-aware helper function for parsing timestamps
Previously we just used `dateutil.parser.parse` to handle timestamps, but
that's not really good enough as we can't be sure whether or not we get a
valid timezone back, and when we're comparing `datetime` objects they must
all be either tz-aware or tz-naive.

This commit implements a new helper function that parses timestamps using
`dateutil.parser.parse` and then either makes sure that the resulting object
*does* or *does not* include a `tzinfo`.

This fixes a bug where the timestamp parsed from a basebackup's
`backup_label` wasn't recognized and we ended up with a naive `datetime`
which later on would cause errors when it was being compared to other
tz-aware `datetime` objects.