restore: Allow using external tar command
Python's tar library performs badly when running in the main process
that does decryption and decompression. It is also otherwise slower than
native tar application.

Switch from the earlier model of creating temporary file and pulling
data from there to Python's tarfile library to a push model where data
is directly pushed through all transformations. This allows passing the
data directly to external tar application, which can do the processing
faster than the Python library and allows skipping the creation of any
temporary files completely. Python's tarfile based implementation is
provided as an alternative for native tar for environments where GNU
tar is not available or using it is undesirable.