object storage: refactor iteration methods to include prefixes
Refactor the `list_iter` method into `iter_key` that yields both objects
(as `list_iter` does) and prefixes. For `deep=False` (the default), only
objects that are immediate children of the key are yielded, while for
objects further down the hierarchy their prefixes. For `deep=True`, all
objects with the given key prefix are yielded. Note that in this case
there are no prefix items yielded.

The yielded items are instances of `IterKeyItem` , which is a
`namedtuple` with the `type` attribute set to `"object"` or `"prefix"`.
The `value` attribute is a `dict` for `"object"` type items, and an
`str` for `"prefix"` types.

`BaseTransfer.list_iter` and `BaseTransfer.iter_prefixes` filter over
`list_iter` items yielding only child values of the relevant type.

Backwards compatibility is maintained for `list_iter` and `list_path`.