Filter Metadata's keys from S3.
Minio S3 implementation return metadata with uppercase letters ( first ones ), triggering a bug in pghoard that cannot find the key ( KeyError ).
This simple patch simply always convert the keys of the metadata dict to lower case, solving it all .
Also, my previous patch did not filtered ALL the metadata, some of the
pghoard part was working, but some execution path did not. That is why
i am removing it as it is now useless