Keep track of WAL SHA1 hashes to allow verifying contents
Previously archive sync only verified file existence in remote file
storage but in some cases the files may be present but invalid. E.g.
in case old master is replaced because it became unresponsive the node
might still be running but just very slowly and it could manage to
upload WAL file for old timeline that is invalid but new master didn't
replace the file with a valid one because it was happy with the fact the
file exists even though it wasn't usable.

SHA1 was used due to its availability in Python's built-in hashlib. Some
other hashes might be more suitable due to better throughput but even
for very busy environments the performance of SHA1 should be good enough
not to cause noticeable extra load so the extra overhead of using hash
that doesn't ship with Python isn't worthwhile.