pghoard: Azure upload uses tell method
At least when using azure-storage-blob==2.1.0, Azure upload (when done
in default parallel=2 mode, with seekable object), failed without
proper tell method on file object as the faked None was not equal to 0
when upload started.

Left the 'tell' wrapper in place for where it probably works (seekable).
Azure code chooses which path to use based on seekability of the file
object, as well as presence of tell+seek methods, so adding tell does
not in and of itself change path which is used.

Added also unit test which covers this case - BytesIO is seekable.