Add support for delta basebackups
Delta backups are based on the hashes of the files. The idea is to
create a Snapshot of the data files, which results in a manifest file,
describing the hashes of all the files needed to be backed up.
New hashes are uploaded to the storage and used together with complementary
manifest from control files for restoration.
In order to deal with situation when hash changes during the upload,
the file is prepared for upload with a temporary name first and then being uploaded
to the name with the real hash.