rohmu/google: HttpError when streaming perfectly aligned file
When doing streaming upload, a `*` can be used for the total file size
in the `Content-Range` header of intermediate chunks. But the total
file size should be provided in the final chunk so that the uploading
can be finalized. Add a `peek` method into `MediaStreamUpload` so that
we can return the total file size when reading the fianl chunk.