Refactor queue events.
Previously, queue events were dict, with different keys meaning
different things.

Refactor this to use specific events for each type of queues. The main
benefit is to standardize the use of pathes, and not have every module
(compressor, transfer) responsible for guessing the file destination
based on it's name or path or whatever. This will allow us to handle
more easily different types of files as proposed in the tentative v3
backup format, and to implement more steps in the compression / transfer
pipeline.

The actual file data is now stored in a source_data field, which can be
a Path or a BytesIO. This will allow us to use different temp storage
strategies more easily too (TemporarySpoolFile, NamedPipe, BytesIO,
NamedTemporaryFile...).