Add possibility to filter small files when doing delta backup
Sometimes it's useful to not treat small files as separate delta backup
hash files, rather consider them a part of bundle which can be managed
separately. E.g. in cases where we don't want to upload such files to the
object storage separately, rather upload only big enough files, but the
small ones put together and upload as an archived bundle. We don't need to
bother calculating hash in this case even. We could have done the same
for embed files as well, but this should stay for backward compatibility.