teach rohmu to notifiy when it writes anything [AD-756]
This allows robjecter to be informed everyime pghoard (rohmu) writes
anything. This can be controlled with the presence of extra config
passed to the constructors of each object storage driver.

The plain_key is needed as robjecter itself uses rohmu so the prefix
gets prefixed twice.
size is not strictly needed but helps make robjecter's work easier. And
here size is easily avaialble statically or dynamically (in case of
multipart-uploads). In case of multipart uploads the size could
calculated using the callback mechanisms locally however each driver
reports callback in a different manner and in some cases the callback
might not be called if a multipart upload was not necessary.

[AD-756]