Support for atomically creating a file with content
Many functionalities in rohmu end up being broken if anything "bad"
happens during a function call, since a file can be half-written.
We cannot perform the usual write-to-temp-then-move-to-final in many
places, since we've got no control on the original source directory,
which could be used by client code for other purposes.

So, in order not to need to change too many interfaces, we leverage
Linux O_TMPFILE in order to create an anonymous temporary file in the
same directory and hence same filesystem as the one that was requested,
and we atomically create a real hardlink on the fs only if the operation
completely succeeded.

Implementation is taken from Linux open(2) manual about O_TMPFILE, almost
verbatim.