Make sure that calls to get_contents_to_file() are actually atomic
In some implementations, calls to get_contents_to_file() weren't
actually atomic; it could happen that a file was left hanging if the
process was killed, or that temporarily-named files were created in the
same directory as the destination and left hanging on errors.

We use our new atomic-open-for-writing implementation that should fix
all these issues.