notifier: fix notifications for some storage types, add tests
For Azure, Swift and SFTP object storages, the attempt to get object size
for completion notifications was incorrect, calling `os.path.getsize()`
with a file-like object, instead of a path.

Given that the object to store may not be a file on disk, that can be
seeked or `stat()`ed, we need some other way to obtain the number
of bytes.

For these storages, the store call doesn't seem to return the object
size in a convenient way, so make use of the progress callback we
implement for object storage backends. That gives us the number of
bytes transferred, that we can then pass on to notifier.

For tests, start with these three backends that seemed to have been
broken by #67. Some libraries used by them are not in our
requirements.dev.txt, so settle for mocking those imports.