notifier: sanitize metadata before trying to json dump
this is needed because metdata can contain complex objects like
datetime

An easier way to fix this would be to move sanitize_metadata out of
baseclass to utils and directly call it from notifer, but in case of
azure the hypen(-) in keys needs to be replaced with underscore (_).
and notifer has no idea what storage driver was is use and if hyphen
replacement should be done.