metrics for s3 operations
this is very adhoc approach. the self.stats.operation calls are
sprinkled everywhere there is s3client being used. The operations are
not directly mapped to actualy s3 categories in code, but can be done
through config, so that remapping can be done easily as the business
requirements change.

the notifier interface could have been extended but the interface is
very limited in functionality plus notifier does not care about iter_key
or object_downloaded kind of notifications

dont use a decorator to report storage_operations for s3

pro - this is cleaner approach
con - decorator assumes the "self" parameter which looks slightly ugly
    - if extra tags are needed, things become weird

operations are reported even if there was some kind of error, eg: an
auth error while iterating will still report as iter_key operation even
though that op never actuall took place. This is not necessarily bad as
the API call was made and not sure how server prices such situation

add basic test for metrics report on s3