metrics for google's storage
we could assume that retries are free and not count those, then passing
`operation` all the way down to `retry_on_reset` need not be necessary.
and just a metric at the end would suffice, but we might end up
under-counting if there are issues uploading huge sized files.

But if we do care then making an assumption about the size of data being
retried to be *_CHUNK_SIZE is not a bad idea. As we might end up
over-counting, but that is not necessarily  a bad idea