Fix S3 delete_keys harder
Apparently the fixed code wasn't right either but the tests I ran on
it outside were insufficient (they did not notice delete failures due
to prefix being omitted as we don't parse the delete_keys response).

Now actually added unit tests for the s3 delete_key* and also added
prefix parameter to s3 storage tests in general to ensure it is passed
along to the API calls that we check.