rohmu,typing: disallow_incomplete_defs = true
In s3.py inside the `multipart_upload_file_object` we should really use
the incremental progress function. The size parameter is unused so we
ignore the fact that it can be `None` instead of an `int`.

In swift.py the `content_length` seems to possibly be `None`. Don't
report progress if it is `None`.