transfer-pool: implemented transfer-pool
The new rohmu.transfer_pool module provides an API to easily define a
transfer pool that can be used by multiple threads to safely share
transfer instances.

It is possible for transfer instances to mark themselves as threas-safe,
in which case the same instance is used by all of the threads using a
pool.

The thread-pool has configurable age for the transfers, allowing to
expire old ones.

The pool can be used both as a context manager or via a
`get_transfer` method, in which case the user must remember to call
`return_to_pool`. The `get_transfer` methods returns a proxy that should
help the user detecting bugs by forbidding method calls after the `return_to_pool`
method is called.