s3: do not treat an empty iter_key response as successful
When requesting a list of files using the S3 protocol, a valid
response will always include a Contents key in the JSON-encoded body.

If this is not present, allow an exception to bubble up, instead of
pretending that the bucket/prefix is empty. It is preferable to raise
an exception than to pretend to have a valid response.