s3: do not mask failures to retrieve metadata
If a file's metadata is being requested and it cannot be found, this
is a real error. We should allow this exception to bubble up, rather
than masking it.
Some backends (e.g. minio) may provide a `KeyCount` value. If this
is present and zero, assume no records are present, and do not
look for the `Contents` list.