Add option to preserve trailing slash on deletes
Most object storages support having a slash in the object
name and rohmu should be able to delete such objects.

Let the default behaviour the same but introduce an option
`preserve_trailing_slash` which allows to control whether the trailing
slash in the key are stripped or not.

Added test coverage for `delete_key` and `delete_keys` methods for all
transfer types (several had no coverage of it).