s3: respect max num of parts per upload
maximum number of parts per upload for s3 compatible
api is 10_000.

make sure we respect that and if needed we increase
the chunk size to avoid starting a multipart upload
with number of chunks (parts) that we already know
is not valid.

when parts number exceeds 10_000 parts, chunk size
will be increased to fit file size in exactly 10_000
parts.