Handle SFTP disconnections better
The SFTPTransfer class is rather naive in its treatment of SFTP
connections; if the server ever disconnects due to, for example, network
connection issues or rebooting for maintenance, the SFTPTransfer will
enter a permanently failed state even if the server is back up.

This change introduces handling for socket closed errors. Idempotent
operations are retried once with a new connection. And for all
operations, if they fail (even after a retry) then the original
"Socket is closed" exception is still raised to the caller, but the
current client & connection is invalidated so the next usage of the
client will retry connecting.