s3: new option to set min_multipart_chunk_size
this would be helpful to patch a running system
to use bigger chunk size.

for example if the system is running on a machine
with 4GB RAM and has ~9 MB default chunk size,
uploading a big stream of 150 GB, but the size
might be unknown during the upload for the stream,
by setting the new option min_multipart_chunk_size to
~20 MB we can successfully upload the stream.

min_multipart_chunk_size is optional, and when set
the max between min_multipart_chunk_size and default
chunk size, which is calculated based on available
memory will be used.