s3: support OVH EU S3-compatible storage
The AWS S3 documentation for `HeadBucket` specifies:

> If the bucket doesn't exist or you don't have permission to access it,
> the `HEAD` request returns a generic `400 Bad Request`, `403 Forbidden`,
> or `404 Not Found` HTTP status code.
> A message body isn't included, so you can't determine the exception
> beyond these HTTP response codes.

Which is unfortunate since it is often the case that buckets has
`s3:ListBucket` permission set with a condition to only allow listing
under certain prefixes, which causes `HeadBucket` to fail.

This hasn't been a problem before because generally most S3-compatible
storages will return `403 Forbidden` for a bucket that exists but for
which we do not have `s3:ListBucket` permission or that permission is
only allowed under a prefix.

We recently discovered that OVH EU S3-compatible Object Storage service
instead returns a `400 Bad Request` which causes rohmu to assume the
bucket doesn't exist and than the `CreateBucket` will fail with
`BucketAlreadyOwnedByYou`.

So make sure that if we see that specific error code we do nto bail out
and instead continue on since everything is actually good. We do not
consider `BucketAlreadyExists` because that could mean the bucket exists
on a different account which may be an error in the configuration.

Note that the behaviour for `create_if_needed=False` is unchanged and
hence wrong for OVH EU S3-compatible storage. We will address it in the
future likely falling back to `GetBucketLocation` and/or `GetBucketAcl`
to determine if the bucket exists but we don't have the correct
permissions for `HeadBucket`.